refactor(config) : global 디렉토리 구조 변경
