+++ src/main/java/com/sns255/sns255/domain/board/entity/Board.java
@@ -1,6 +1,6 @@
 package com.sns255.sns255.domain.board.entity;
 
-import com.sns255.sns255.global.entity.BaseEntity;
+import com.sns255.sns255.global.config.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Getter;
